Created columns in the magazine detail component which display accepted and rejected stories respectively
